add job_id to all logger file entries for submit dmf determination
